requirements.txt
Django==2.1.5
django-model-utils==4.0.0
djangorestframework==3.8.2
